+++ packages/core/chain/tx/broadcast/resolvers/evm.ts
@@ -25,11 +25,7 @@ export const broadcastEvmTx: BroadcastTxResolver<EvmChain> = async ({
     !isInError(
       error,
       'already known',
-      'transaction is temporarily banned',
-      'nonce too low',
       'transaction already exists',
-      'future transaction tries to replace pending',
-      'could not replace existing tx',
       'tx already in mempool'
     )
   ) {